Support multiple RabbitMQ broker URLs
`kombu` (already used by `celery`) is a higher-level messaging library
than `amqp`.  For our purposes, we can treat it mostly as a wrapper for
`amqp` with some slightly more convenient interfaces, but it has one key
feature that's of interest to us: it supports multiple RabbitMQ broker
URLs with round-robin failover between them.  This makes it possible to
configure Launchpad to use RabbitMQ with high availability: if we have a
RabbitMQ cluster, then we can configure Launchpad with broker URLs for
all the nodes in the cluster, and if one fails then `kombu` will
automatically fail over to the next.

To make it practical to configure this, I had to add a
`rabbitmq.broker_urls` configuration option, which supersedes the
existing broken-out configuration options (`rabbitmq.host`,
`rabbitmq.userid`, `rabbitmq.password`, and `rabbitmq.virtual_host`).
For backward compatibility, the old options continue to work as long as
`rabbitmq.broker_urls` is unset.

This also includes upgrading to oops-amqp 0.2.0, since that includes a
change to accept connection factories that return `kombu` connections.