Fix json.dumps call for Python >= 3.6
The `json.dumps` call in `lp.bugs.model.apportjob` started failing on
Python >= 3.6 after the change to drop the dependency on `simplejson`,
because `json.dumps` changed in 3.6 to make most arguments keyword-only.
Passing the encoding as a positional argument has probably been
incorrect for a long time since the encoding wasn't the next argument
after the object to dump, but this hasn't mattered since Python 3
anyway.