Add a socket timeout for librarian client connections
The internal librarian client creates a socket with no timeout, so if
anything goes wrong with the connection then it only recovers by being
killed manually.  Add a socket timeout so that it gives up eventually.
(Note that this timeout applies to individual operations on the socket,
not to uploading or downloading entire files.)

The default timeout is deliberately quite generous to minimize the
chance of causing failures due to connections being merely slow rather
than stuck, although we may still find we need to tune it.  This isn't
intended to ensure that librarian requests fit within timeout budgets
set for web requests or similar, but just as a backstop to ensure that
we don't get stuck waiting forever.