Allow authenticating to git using access tokens
We added personal access tokens last year for some initial webservice
API use cases, and with the intention of eventually extending them for
git authentication as well, but never quite got round to the latter.
This commit adds a new `repository:pull` scope to go with the
previously-unimplemented `repository:push` scope, and implements both of
those for git authentication over HTTPS.

The way that the git XML-RPC endpoint works means that we have to do
this in a few different places: the initial entry point for HTTPS is
`authenticateWithPassword`, which just verifies that the username and
access token match and returns some information about them which turnip
sends along with later calls.  Those later calls (`translatePath`,
`checkRefPermissions`, etc., which can only be called after successful
authentication) then each need to handle more specific checks on the
access token, such as whether it's for the correct repository and
whether it permits pulling or pushing as appropriate.  Fortunately, the
same general idea was already in place for macaroon authentication, so
for the most part we can just extend that.

This introduces the novel-to-Launchpad feature of a token that permits
pushing but not pulling.  While I don't have a specific use case for
this, I didn't think it was wise to just have any token allow pulling
(since the `repository:build_status` scope currently only allows using
some specific webservice API endpoints), and it seemed clearer to add a
specific scope than to have `repository:push` imply `repository:pull`.
I've made sure that this works correctly with a real git client.

Once this is deployed to production, we can switch `snapcraft
remote-build` over to use this instead of macaroons, and then begin the
process of deprecating user macaroons for git.

LP: #1995535