Return proxied objects from makeGitRefs
`LaunchpadObjectFactory` issues `UnproxiedFactoryMethodWarning` when its
methods return objects not wrapped in a security proxy, since that tends
to result in tests that are less accurate simulations of production.

This required tightening up a number of tests.  Also,
`GitRefMixin.__eq__` and `GitRefMixin.__hash__` need to be more careful
to cope with the `commit_sha1` attribute of `GitRefDefault` objects
raising `NotFoundError`.