Close requests sessions more promptly
This fixes a number of particularly obscure-looking `ResourceWarning`s
from the test suite that look something like this:

  ResourceWarning: unclosed <socket.socket fd=8, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=6, laddr=('127.0.0.1', 46726), raddr=('127.0.0.1', 39003)>

(I finally got curious enough about this to put a breakpoint at the top
of `lp.scripts.utilities.warninghandler.launchpad_showwarning` so that I
could track it down.)

I haven't fixed the case in `ArtifactoryPool` because the disk pool API
makes that difficult and will need more extensive refactoring; I've just
left an XXX comment for now.

This looks long, but it's almost entirely reindentation due to using
sessions as context managers.