Add API/UI for setting builder constraints on Git repositories
Using a `DelimitedListWidget` for this is pretty ugly (it just gives you
a text area and splits what you enter on whitespace), but it's
functional, and this is only visible to admins anyway.

I believe the `launchpad.Admin` permission on repositories was
previously unused, so I repurposed it for this since some builder
resources will be restricted.