Bump default librarian client socket timeout to 15 minutes
It turns out that what I thought was a generous timeout of a minute
causes problems for some scripts that upload large files, probably
because the librarian storage protocol doesn't acknowledge a block at a
time after the initial header but has to store the entire file first.  I
didn't expect that to take more than a minute, but that's what we're
seeing.  Since this was only ever intended as an improvement over
waiting forever, bump the timeout to 15 minutes.

LP: #1998037