Switch to Celery's RPC result backend
Using Celery's AMQP result backend produces the following deprecation
warning:

  The AMQP result backend is scheduled for deprecation in version 4.0
  and removal in version v5.0.  Please use RPC backend or a persistent
  backend.

The RPC backend scales better due to using a results queue per client
rather than a results queue per task call, but it only works if the
process that initiates the task is always the process to retrieve the
result.  This is fine for Launchpad, though: we generally persist the
results of jobs in our database instead.  Our job runner ignores Celery
task results unless `BaseRunnableJob.celery_responses` is set, and
that's only the case in tests where the process initiating the task is
indeed also the one that retrieves the result.