Disallow ProductSeries translation exports to most users
Exporting translations for an entire `ProductSeries` can be benign, but
in some cases (e.g. `/linuxmint/latest` on production) it can be very
expensive; and users who aren't admins or responsible for the project in
question should generally export an individual `POTemplate` or `POFile`
instead.  Restrict this operation to admins, owners of the project,
release managers of the project series, or owners of applicable
translation groups, in much the same way as we already do for
translation exports of entire source packages.

LP: #1998037