Fix private team visibility queries for some users
`PublicOrPrivateTeamsExistence.checkAuthenticated` was slow for users
with archive subscriptions via many different teams, because it
performed security checks on all those teams when computing the set of
archive IDs to which the user has access via subscriptions in order to
see whether any of them match one of the target team's private PPAs.  We
can safely skip those security checks, because we know that the user can
view all their own subscriptions.