Strip trailing backslashes from Artifactory property values
I've previously established empirically that backslashes must not be
quoted when setting Artifactory properties (contrary to the public
documentation), and the `encode_properties` function used internally
when setting properties behaves as follows:

    >>> from collections import OrderedDict
    >>> from artifactory import encode_properties
    >>> properties = OrderedDict([("key1", "\\;=|,\\"), ("key2", "value")])
    >>> print(encode_properties(properties))
    key1=\;\=\|\,\;key2=value

This encoding clearly leaves no way to separate the fields
unambiguously, and the result is that ";key2=value" ends up being
appended to the value of the "key1" property, which is no good.

Stripping trailing backslashes is a horrible workaround, but I don't see
a way around it given how the Artifactory API works.  I expect trailing
backslashes in property values to be extremely rare in practice, and
they'll only show up in informational properties auto-populated by
Artifactory itself, not by anything functional.