Refactor _updateByHash to clarify logic
William Grant pointed out in
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/373971
that the flow of `Publisher._updateByHash` was quite difficult to
understand, and proposed an alternative structure in
https://code.launchpad.net/~wgrant/launchpad/+git/launchpad/+merge/390811.
Over two years later, I've finally got round to digesting and polishing
this.

The basic idea is that, rather than interleaving database/disk queries
and actions, we now gather all the information we need from the database
and from the archive on disk at the start, and then separately take all
the actions needed to reconcile them and to keep `by-hash` directories
up to date.

No functional change is intended here.