Turn ArchiveFile into a history table
This adds `date_created` and `date_superseded` columns, which we'll be
able to use later to query for files that existed in an archive at a
given point in time.

The publisher changes a little as a result of this, mostly simplifying
it: unscheduling `ArchiveFile`s for deletion would now require creating
new rows in order to keep the `date_created`/`date_superseded` windows
intact, so we now just follow the normal code path to create new files
in that case rather than bothering with the previous optimization.

LP: #1765933