Fix BranchDeletionView's checks on deletion requirements
The recent work to allow registry experts to delete branches involved
changing the permission for deleting branches from `launchpad.Edit` to
`launchpad.Moderate`.  That work also changed
`BranchDeletionView.display_deletion_requirements` to check for the
`launchpad.Moderate` permission on deletion requirements (other objects
that may need to be altered or deleted in order for the branch deletion
to proceed).  However, those other objects are always of other types
(e.g. merge proposals or recipes), and so that part of the change was
incorrect: we should still check `launchpad.Edit` here.

Fixes a test failure in
`lib/lp/code/stories/branches/xx-branch-deletion.rst`.