Add type annotations for arguments sent to builders
It's difficult to do a completely accurate job of this with Python 3.5,
but we can get most of the benefit (in particular, protection against
simple typos and against passing the wrong type for an argument) from a
single `TypedDict`.  The annotation is long and complicated because the
protocol has evolved in quite an ad-hoc way for many years, and a number
of the arguments were difficult to describe concisely without context;
still, the exercise of going through the whole thing was quite a useful
one and found a few forgotten deprecations to clean up.

There should be no functional change from this commit.