Drop missing-file warning in Publisher._getCurrentFiles
As part of the recent refactoring in commit
82dfb163d64512e0d95af290adbe9b7bade70709, I added a warning if an entry
in an archive's Release file doesn't exist on disk.  However, this
actually turns out to be routine and expected; QA testing reminded me
that Release files include entries for uncompressed versions of Packages
and Sources files for the benefit of clients, even though only the
compressed versions exist on disk.  Launchpad scripts set up a handler
that logs `WARNING` and higher log messages as OOPSes, so this warning
ends up being unreasonably noisy.

Since the warning is recent and has never been deployed to production,
it's safe to just drop it.