publish-distro: Add an --archive option
This fixes an annoyance when performing QA: `publish-distro` already had
`--ppas` options and similar to tell it to run over all archives of a
particular kind, but previously the only way to get it to run over a
single archive was to hack the `is_ppa_public` etc. functions in place.
This option should make publisher QA a little more convenient in future.