Temporarily disable librarian.client_socket_timeout
For some reason this timeout is causing the translation export script to
fail to upload exported files to the librarian much more often than
usual.  I'd still like to track this down properly so that we can get
the robustness improvement from this timeout, but it isn't worth causing
major user-visible issues for that.

LP: #1998616