Return proxied objects from all remaining factory methods
`LaunchpadObjectFactory` formerly issued `UnproxiedFactoryMethodWarning`
when its methods returned objects not wrapped in a security proxy, since
that tended to result in tests that are less accurate simulations of
production.  This commit finishes ensuring that all factory methods
return proxied objects, and converts that warning into an error so that
we can't accidentally reintroduce this problem.

This required tightening up a few more tests to log in or to remove
security proxies before accessing proxied objects.