Fix query plan for specification searches
When given base clauses, `search_specifications` is intended to assume
that those are more selective than other criteria: for example, starting
by filtering specifications by project is likely to cut down the row
count a lot more than filtering by status.  However, it constructed a
query of the form:

    WITH RelevantSpecification AS (...)
    SELECT Specification.*
    FROM Specification
    WHERE
        Specification.id IN (
            SELECT RelevantSpecification.id FROM RelevantSpecification
        )
        AND ...;

This caused the PostgreSQL planner to find relevant specifications using
the base clauses, then filter all specifications according to the other
clauses, then join the results of those two searches together: the worst
of both worlds.  Instead, we want a query of the form:

    WITH RelevantSpecification AS (...)
    SELECT Specification.*
    FROM
        Specification
        JOIN RelevantSpecification
             ON Specification.id = RelevantSpecification.id
    WHERE ...;

This results in a query plan that finds relevant specifications using
the base clauses and then filters the result of that search using the
other clauses.  Much more sensible, and takes milliseconds rather than
seconds.

LP: #1995686