Make publisher ignore pending-deletion Artifactory archives
Actually deleting these is unlikely to be at all urgent and can be done
manually if needed.  Meanwhile, the previous code was causing the
publisher to crash with `NotImplementedError: Don't know how to delete
archives published using Artifactory`, which caused anything after that
point in the publisher's to-do list not to happen.

I also fixed a couple of existing test assertions that I don't believe
were quite doing what the test author intended, since they would have
passed regardless of whether the archive in question was deleted.