Fix test file names in DSCFileTests
The `.dsc` needs to be matched by
`lp.archiveuploader.utils.re_issource`, as otherwise
`determine_source_file_type` would return None, which isn't valid for
`SourcePackageReleaseFile.filetype`.

This would have failed if the `INSERT` statement had been flushed to the
database since the database column is already `NOT NULL`, but as part of
the conversion of `SourcePackageReleaseFile` to Storm I tightened up
some of the ORM column definitions a little, which means that it now
fails slightly earlier.