Suppress timeline in PoolFileOverwriteError OOPSes
The traceback for `PoolFileOverwriteError` (raised when the publisher
attempts to overwrite an existing file in a pool with different content)
tells us everything we need to know: we get both of the checksums and
the target path.

However, the OOPSes also contain an enormous timeline mainly consisting
of every SQL statement the publisher has issued as part of processing
that archive.  These timelines tend to cause the OOPS processing system
to back up, and provide essentially no information of any value.
Instead, just arrange for these OOPSes to be raised with a temporary
empty timeline.