Retain more ArchiveFile history
Source and binary packages are removed from the published archive on
disk a day after they are no longer referenced by current index files;
package files in the librarian are typically expired a week after all
their associated publishing history rows have been marked as removed
from disk, although files in some archives (primary archives and certain
named PPAs) never expire.

Archive index files, such as `Packages`, `Sources`, and `Release` files,
currently work quite differently: old versions of these files are
published in `by-hash` directories on disk so that clients can continue
to make use of them for a grace period of a day, after which they are
both removed from disk and removed from the database.

The latter approach is problematic for the forthcoming snapshot service,
because it means we can only serve snapshots of index files for as long
as they are still published on disk: not much more useful than what we
already have!

This commit implements a much more useful policy.  We now treat archive
index files similarly to source and binary packages: they're removed
from disk after a day as before, to prevent `by-hash` directories
getting unreasonably large and causing bloat on mirrors, but now we
merely set the `ArchiveFile.date_removed` column rather than deleting
the row entirely.  The `expire-archive-files` script now handles
`ArchiveFile` as well as `SourcePackageReleaseFile` and
`BinaryPackageFile` rows, which means that old snapshots of the primary
archive and some other important archives will be retained in the
database and the librarian.

This will of course cause the librarian to grow more quickly, but some
back-of-the-envelope calculations suggest that even if we retain old
archive index files in perpetuity then this will be on the order of an
extra terabyte per year, which seems quite manageable.