Make BranchSet.getByPath forbid paths with suffixes
It previously accepted suffixes and discarded them.  This stricter
approach is consistent with how `GitRepositorySet.getByPath` works, and
it makes things easier for Breezy (which already walks up the path
structure until it finds a match, but we were deceiving it by returning
inexact matches).  It's also a better match to the existing API
documentation, which didn't mention anything about accepting suffixes.

LP: #2003950