Ignore authentication for git push notifications
Back when Ioana was implementing sending pack statistics for git push
notifications (which are emitted by turnip at the end of each push), one
of my review comments was that, since this was directly updating
properties of the repository, the call ought to be authenticated.  I
remember being slightly ambivalent about that even at the time, but we
went along with it since it seemed a relatively harmless piece of
caution.

However, recently I've been experimenting with getting snapcraft to use
personal access tokens, and one thing I noticed was that authenticated
push notifications make it difficult to use appropriate expiry times.
`snapcraft remote-build` creates a repository, generates an access token
for it, and then pushes to it; the token relieves the user of the need
to set up an SSH key before this system can work.  The time between
generating the token and starting the push is very short, and the token
is only used once, so it seems as though a short expiry time (perhaps a
minute) should be more than enough.  However, because `notify` is called
at the end of the push and also checks authorization, we have to choose
an expiry time long enough to push an entire potentially-large
repository over a connection of unknown speed.

We already have to trust turnip to handle part of the git authorization
process, so an authorization check that effectively just ensures that
turnip hasn't processed a push without appropriate authorization doesn't
add any interesting security.  Being able to update pack statistics on a
repository would be a very low-value attack in any case.  On the other
hand, being able to use short-expiry tokens effectively feels like a
much clearer security advantage.

I'm therefore reversing my previous view: it seems acceptably safe to
skip authentication for `notify`, given the other safeguards around that
and the benefits for use of personal access tokens.