Move ZCML authorizations directive to lp namespace
While experimenting with https://github.com/collective/zpretty, I
noticed that Launchpad installs a number of ZCML directives under the
XML namespaces http://namespaces.zope.org/browser,
http://namespaces.zope.org/mail, and http://namespaces.zope.org/zope.
As well as feeling a bit rude, this also makes it difficult to integrate
properly with auto-formatting tools such as `zpretty` that can benefit
from being told about details of the preferred ordering of attributes
for each tag.

We have a perfectly good http://namespaces.canonical.com/lp XML
namespace that we already use for a few ZCML directives, so I think it
would make sense to move our other local directives to that.  There'll
be a bit of readjustment, but it should also have the benefit that it
will be more obvious at a glance where to start looking for the
definition of a given directive (this is something I've had difficulty
with in the past).

As an initial proof of concept, this commit moves the `authorizations`
directive to http://namespaces.canonical.com/lp.