Move ZCML securedutility directive to lp namespace
As with commit 57196c1d6a220f00039e027768cc05368135f556, it makes sense
to move our local `securedutility` directive to our own namespace rather
than poking it into the Zope namespace, even though it's a bit like the
Zope `utility` directive.  This is a big patch, but it's mostly
mechanical.