Fix ArchiveAPI tests to not create bad ArchiveFile combinations
DB patch 2211-15-0 introduces a unique index to ensure that there may be
only one non-superseded `ArchiveFile` row for a given path in a given
archive at any one time.  The `ArchiveAPI` tests unintentionally
violated this.

This situation shouldn't happen on production, and I've checked that
there are no violations in the most recent production dump that was
loaded into staging.