Avoid substitution issue with next_url and cancel_url
Launchpad's form views are allowed to define `next_url` and `cancel_url`
to indicate where the form should redirect to after a successful
submission or if the user chooses to cancel.  Some do this by assigning
to those attributes (especially if the URL is computed based on the
submitted form), while some do this by defining those attributes as
properties (especially if there are multiple form actions that use the
same URLs).  There's no obvious way to settle on one approach or the
other here, but until now it hasn't mattered.

Unfortunately, `mypy` detects a technical difficulty here.  If a base
class declares something as a plain attribute, then a subclass that
redefines that as a property without also providing a setter (which
wouldn't really make sense here) violates the Liskov substitution
principle, because assigning to that attribute is allowed for instances
of the base class but forbidden for instances of the subclass.

However, if we change the base class to make those attributes be
read-only properties (which is all that the general form view
infrastructure needs), and then redefine them as plain class attributes
(e.g. with `next_url = None`) in all the subclasses that need to assign
to them at some point, then the type checker is happier and everything
still works.  This slightly increases the friction of writing views that
assign to those attributes, but for many cases of new views properties
will be just fine, and it's lightweight enough not to be too big a
problem anyway.