Grant Launchpad devs selected edit permissions for PPAs
There are a couple of common requests for the Launchpad team regarding
PPAs, e.g. increasing size, set them to private, or configure them to be
used with Artifactory.

Due to historic reasons, this was only possible for Colin (or William).

There is no reason why all Launchpad devs should not be able to process
these requests.

As a starter the permission have been update for API access.

Later on we also want the changes to be made via UI.