Stages now depend on architecture
Stages now is a dictionary where each key corresponds
to the architecture and the value is the old stages array.
Thus we'll have number of build farm jobs equals to the
number of architectures.
If an arch is present in the pipeline but missing in previous
stages an error is raised.

LP: #1999591