Allow archive subscribers to use +sourcefiles and +files
This will be needed by the `debuginfod` service, but it seems reasonable
anyway.  It has the effect that subscribers will be able to see
historical files in archives as well as current ones; but they can build
much of that information up over time anyway, so this was never a robust
boundary, and the in-progress archive snapshot service will soon allow
equivalent access.