Rename "blacklist" to "blocklist" for DistroSeriesDifferences
Normally we'd need to take a bit more care than this, because renaming
HTML fields means that users may have problems if they load a form
before the upgrade and attempt to submit it after the upgrade.  However,
this UI machinery is sufficiently rarely used that I don't think it's
worth bothering with anything more careful.