Fix datetime handling in ArchiveAPI.translatePath
Python's XML-RPC code doesn't handle `datetime` objects quite the way I
thought it did, and due to a deficiency in `zope.publisher.xmlrpc` we
don't currently have a way to opt into cleaner behaviour, so we have to
cope with an awkward conversion.

The tests now do proper argument serialization and deserialization so
that we don't miss other similar issues, which also required coping with
the fact that XML-RPC doesn't preserve the microsecond part of
`datetime` objects.