Include traceback supplements and info in timelines
Some code adds extra information to tracebacks (documented in
https://zopeexceptions.readthedocs.io/en/latest/narr.html): in
particular, the page templating engine does this, and it makes it much
easier to work out what tracebacks in the middle of page rendering mean.
Unfortunately, this information wasn't recorded in the individual
entries in an OOPS timeline, which in some cases made OOPSes very hard
to decipher (LP: #2009357 is a case in point).

Adjust the way we construct `Timeline` objects so that they format
tracebacks using the machinery in `zope.exceptions`, which knows how to
deal with this extra information.