charm: Fetch dependencies using charmcraft
Now that `ols-layers` is open-source, we can fetch it directly using
`charmcraft`, which is much simpler and should let us use Launchpad
charm recipes.

I had to split out `charm/layer/` to a separate
https://git.launchpad.net/launchpad-layers project, because `charmcraft`
can't deal with referencing the parent directory from one of a charm's
parts.  That separation isn't ideal, but the benefits of being able to
use charm recipes should outweigh the costs.