charm: Add a simple launchpad-admin charm
This is useful as a way to get hold of database superuser credentials in
a charmed deployment, without having to SSH into a `postgresql` unit
directly.

For now this is just for interactive use, but I can think of a couple of
natural future extensions: we could do with an action to initialize a
new Launchpad database, and it might well make sense to use this charm
as a place to run our schema upgrade machinery.