Add debbugs-sync script
This adopts an ad-hoc script currently running on a Canonical machine to
sync data from the Debian bug tracking system for use by Launchpad bug
watches.

There's definitely scope for cleaning this up further, adding some kind
of tests, and so on.  For now, I'm working on the theory that it's
better to get production code into more visible and maintainable places
even if it's not in an ideal state, so I've only done some minor cleanup
to pacify `shellcheck` and tidy up logging slightly, as well as
excluding some large files that we don't need in order to reduce load on
both ends of the `rsync`.

See also https://portal.admin.canonical.com/C136072.