Allow initializing a remote database
This is usable from an instance of the `launchpad-admin` charm: it uses
the configuration file to find the necessary database credentials.  As
such, it can be used to set up a Juju-deployed instance of Launchpad
from scratch.

So that I sleep better at night, I added an explicit check that refuses
to destroy the production database this way.