charm: Use rolling restart for launchpad-appserver
As long as we have more than one appserver unit, this ensures that the
application remains responsive during upgrades, using the coordinator
layer to deal with communication between units.

With gunicorn >= 20.0, waiting for the application to be responsive on
each unit turns out to be a matter of setting `preload_app = True`
(which causes the application to be loaded before the arbiter signals
readiness, and also saves a fair amount of memory) and using
`Type=notify` in the systemd service.  I dropped the old soft-restart
mechanism when only `*-lazr.conf` files are changed, since it doesn't
quite work with application preloading and it wasn't all that much of a
win anyway.