Allow editing build_path in charm recipe UI
The charm recipe build protocol supports `build_path`, which changes to
a subdirectory of the checked-out branch before running `charmcraft`.
This is quite useful: for instance, we use it in charms of Launchpad
itself so that we can have multiple charm recipes in the same source
tree that deploy the same project in different ways.

However, it's currently only possible to set this field using the
webservice API, which I think was an oversight from when I first put
together the charm recipe UI.  Fix this oversight.