Defer GnuPG home directory setup to support app preloading
The `launchpad-appserver` charm uses Gunicorn's `preload_app` setting as
part of its arrangements to support rolling restart during upgrades.
However, this causes a problem with the GnuPG handler.  Because it sets
up its temporary GnuPG home directory in `GPGHandler.__init__`, this is
done during ZCML initialization, and when application preloading is
enabled this happens in the Gunicorn parent ("arbiter") process and is
shared between all the child ("worker") processes.  Unfortunately, this
means that when a worker process exits, which can happen during normal
operation, the `atexit` handler that's registered to clean up the
temporary GnuPG home directory is called, thus leaving all other running
workers without a GnuPG home directory.  At this point all GnuPG
operations will fail until the entire service is restarted.

There's certainly work we can do to make the behaviour of our `atexit`
handlers less surprising with child processes in general, but in the
meantime I think it makes sense to ensure that we have a separate GnuPG
home directory for each child process even when using application
preloading; I doubt GnuPG is completely robust against having multiple
processes writing to the same home directory in parallel anyway.  We
achieve this by deferring setup of the GnuPG home directory until we
need to perform an actual GnuPG operation.

I've always been uncomfortable with the way that `GPGHandler` sets
`GNUPGHOME` in the process environment; it seems too much like implicit
action at a distance.  Most of the relevant tools now support passing
this in some other way, either via a GPGME engine parameter or via a
command-line option, so I've switched to that where possible.  There are
a couple of cases where this isn't possible, and in those cases I've
arranged to set the environment variable in a more localized way.