Include built assets in deployment tarballs
We currently serve API documentation, CSS, and JavaScript from our
frontends; at the moment that involves running `make build` on the
frontends to build them.  In order to translate this to Juju
deployments, one fundamental step is to build those assets into the
tarballs we use as payloads for Juju charms.

This adds about 8.7 MiB (9%) to the size of these tarballs, and the
added parts will only be used on one or two Juju applications.  Still, I
think that's an excellent trade-off for the extra flexibility it gives
us, and once everything is converted to Juju it should result in faster
deployments since `make build` is quite slow.