fix-translations-opening: Unapprove existing imports
`scripts/fix-translations-opening.py` is run manually in the case where
opening translations for a new series is interrupted (usually due to
long transactions making it difficult to continue) and leaves
translations in an incomplete state.

The last time we did this, in September 2016, we found that we
additionally had to set the associated translation import queue entries
back from Approved to Needs Review so that
`cronscripts/rosetta-approve-imports.py` could process them again (see
https://pastebin.canonical.com/164459/).  Rather than keeping this
solely in folk memory, let's make the `fix-translations-opening` script
do that as well.

I'm afraid I don't have a good way to test this short of the real-world
run that we're going to need to perform shortly.